refactor: interface to entities
