+++ src/artist/entities/artist.entity.ts
@@ -1,1 +1,5 @@
-export class Artist {}
+export class Artist {
+  id: string;
+  name: string;
+  grammy: boolean;
+}
