+++ src/db/db.service.ts
@@ -1,5 +1,6 @@
 import { Injectable } from '@nestjs/common';
-import { Artist, User } from 'src/types/types';
+import { Artist } from 'src/artist/entities/artist.entity';
+import { User } from 'src/user/entities/user.entity';
 
 @Injectable()
 export class DbService {
